requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.1
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.14.0
djangorestframework-stubs==1.7.0
factory-boy==3.2.1
Faker==15.0.0
filelock==3.8.0
flake8==5.0.4
Flake8-pyproject==1.1.0.post0
gunicorn==20.1.0
identify==2.5.5
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
types-Markdown==3.4.2
types-pytz==2022.2.1.0
types-PyYAML==6.0.12
types-requests==2.28.11
types-urllib3==1.26.25
typing_extensions==4.3.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
